+++ src/views/mini-game/model/usePlinkoForm.ts
@@ -1,18 +1,8 @@
-import { usePathname } from 'next/navigation';
 import { useState } from 'react';
 import { useForm, FieldErrors } from 'react-hook-form';
-import { PlinkoFormType, PlinkoResponse } from '@/shared/types/mini-game';
-import { formatPlinkoData } from './formatPlinkoData';
-import { usePostPlinkoGame } from './usePostPlinkoGame';
+import { PlinkoFormType } from '@/shared/types/mini-game';
 
 export const usePlinkoForm = () => {
-  const pathname = usePathname();
-
-  const match = pathname.match(/\/mini-game\/([^/]+)\/plinko/);
-  const stageId = match ? match[1] : null;
-
-  const { mutate: PostPlinko } = usePostPlinkoGame(Number(stageId));
-
   const { register, handleSubmit, setValue, watch } = useForm<PlinkoFormType>({
     defaultValues: {
       risk: 'LOW',
@@ -26,18 +16,6 @@ export const usePlinkoForm = () => {
 
   const isDisabled = !amount || !risk;
 
-  const onSubmit = (data: PlinkoFormType) => {
-    const formattedData = formatPlinkoData(data, selectedRisk);
-    PostPlinko(formattedData, {
-      onSuccess: (response: PlinkoResponse) => {
-        console.log('게임 결과:', response);
-      },
-      onError: () => {
-        console.error('게임 요청 실패');
-      },
-    });
-  };
-
   const onError = (errors: FieldErrors<PlinkoFormType>) => {
     console.log('제출에러:', errors);
   };
@@ -50,7 +28,6 @@ export const usePlinkoForm = () => {
     selectedRisk,
     setSelectedRisk,
     isDisabled,
-    onSubmit,
     onError,
   };
 };
