+++ src/views/mini-game/plinko/ui/PlinkoPage/index.tsx
@@ -1,84 +1,53 @@
 'use client';
 
-import { useState } from 'react';
-import { PlinkoResponse } from '@/shared/types/mini-game';
+import { usePathname } from 'next/navigation';
+import { useEffect, useState } from 'react';
+import { PlinkoFormType, PlinkoResponse } from '@/shared/types/mini-game';
 import BackPageButton from '@/shared/ui/backPageButton';
 import { cn } from '@/shared/utils/cn';
+import { formatPlinkoData } from '@/views/mini-game/model/formatPlinkoData';
 import { usePlinkoForm } from '@/views/mini-game/model/usePlinkoForm';
+import { usePostPlinkoGame } from '@/views/mini-game/model/usePostPlinkoGame';
 import { PlinkoGame, PlinkoInputBox } from '@/widgets/mini-game';
-import getPlinkoMock0 from '../Mock/getPlinkoMock0';
-import getPlinkoMock1 from '../Mock/getPlinkoMock1';
-import getPlinkoMock10 from '../Mock/getPlinkoMock10';
-import getPlinkoMock11 from '../Mock/getPlinkoMock11';
-import getPlinkoMock12 from '../Mock/getPlinkoMock12';
-import getPlinkoMock13 from '../Mock/getPlinkoMock13';
-import getPlinkoMock14 from '../Mock/getPlinkoMock14';
-import getPlinkoMock15 from '../Mock/getPlinkoMock15';
-import getPlinkoMock2 from '../Mock/getPlinkoMock2';
-import getPlinkoMock3 from '../Mock/getPlinkoMock3';
-import getPlinkoMock4 from '../Mock/getPlinkoMock4';
-import getPlinkoMock5 from '../Mock/getPlinkoMock5';
-import getPlinkoMock6 from '../Mock/getPlinkoMock6';
-import getPlinkoMock7 from '../Mock/getPlinkoMock7';
-import getPlinkoMock8 from '../Mock/getPlinkoMock8';
-import getPlinkoMock9 from '../Mock/getPlinkoMock9';
 
 const PlinkoPage = () => {
-  const [plinkoMock, setPlinkoMock] = useState<PlinkoResponse | null>(null);
+  const [plinkoData, setPlinkoData] = useState<PlinkoResponse | null>(null);
   const [gameRunningCount, setGameRunningCount] = useState<number>(0);
+
+  const pathname = usePathname();
+
+  const match = pathname.match(/\/mini-game\/([^/]+)\/plinko/);
+  const stageId = match ? match[1] : null;
 
   const {
     register,
     handleSubmit,
     isDisabled,
     watch,
-    onSubmit,
     onError,
     setValue,
     selectedRisk,
     setSelectedRisk,
   } = usePlinkoForm();
 
-  const getPlinkoMockData0 = getPlinkoMock0();
-  const getPlinkoMockData1 = getPlinkoMock1();
-  const getPlinkoMockData2 = getPlinkoMock2();
-  const getPlinkoMockData3 = getPlinkoMock3();
-  const getPlinkoMockData4 = getPlinkoMock4();
-  const getPlinkoMockData5 = getPlinkoMock5();
-  const getPlinkoMockData6 = getPlinkoMock6();
-  const getPlinkoMockData7 = getPlinkoMock7();
-  const getPlinkoMockData8 = getPlinkoMock8();
-  const getPlinkoMockData9 = getPlinkoMock9();
-  const getPlinkoMockData10 = getPlinkoMock10();
-  const getPlinkoMockData11 = getPlinkoMock11();
-  const getPlinkoMockData12 = getPlinkoMock12();
-  const getPlinkoMockData13 = getPlinkoMock13();
-  const getPlinkoMockData14 = getPlinkoMock14();
-  const getPlinkoMockData15 = getPlinkoMock15();
+  const { mutate: PostPlinko } = usePostPlinkoGame(Number(stageId));
 
-  const onClickSubmitBtn = () => {
-    const mockDataArray = [
-      getPlinkoMockData0,
-      getPlinkoMockData1,
-      getPlinkoMockData2,
-      getPlinkoMockData3,
-      getPlinkoMockData4,
-      getPlinkoMockData5,
-      getPlinkoMockData6,
-      getPlinkoMockData7,
-      getPlinkoMockData8,
-      getPlinkoMockData9,
-      getPlinkoMockData10,
-      getPlinkoMockData11,
-      getPlinkoMockData12,
-      getPlinkoMockData13,
-      getPlinkoMockData14,
-      getPlinkoMockData15,
-    ];
-    const randomIndex = Math.floor(Math.random() * mockDataArray.length);
-
-    setPlinkoMock(mockDataArray[randomIndex]);
+  const onSubmit = (data: PlinkoFormType) => {
+    const formattedData = formatPlinkoData(data, selectedRisk);
+    PostPlinko(formattedData, {
+      onSuccess: (response: PlinkoResponse) => {
+        console.log('게임 결과:', response);
+        setPlinkoData(response);
+      },
+      onError: () => {
+        console.error('게임 요청 실패');
+      },
+    });
   };
+
+  useEffect(() => {
+    console.log(plinkoData);
+  }, [plinkoData]);
 
   const storedTicketCount = sessionStorage.getItem('ticketCount');
   const ticketCount = storedTicketCount ? JSON.parse(storedTicketCount) : null;
@@ -120,14 +89,13 @@ const PlinkoPage = () => {
               setValue={setValue}
               selectedRisk={selectedRisk}
               setSelectedRisk={setSelectedRisk}
-              onClick={onClickSubmitBtn}
               gameRunningCount={gameRunningCount}
             />
           </form>
 
           <PlinkoGame
             watch={watch}
-            plinkoData={plinkoMock}
+            plinkoData={plinkoData}
             setGameRunningCount={setGameRunningCount}
           />
         </div>
